+++ hubbleutils/data_structures.go
