+++ plugin/evm/gossiper_orders.go
@@ -6,9 +6,9 @@ import (
 	"time"
 
 	"github.com/ava-labs/avalanchego/ids"
+	hu "github.com/ava-labs/subnet-evm/hubbleutils"
 	"github.com/ava-labs/subnet-evm/plugin/evm/message"
 	"github.com/ava-labs/subnet-evm/plugin/evm/orderbook"
-	hu "github.com/ava-labs/subnet-evm/plugin/evm/orderbook/hubbleutils"
 	"github.com/ethereum/go-ethereum/log"
 )
