+++ plugin/evm/orderbook/liquidations_test.go
@@ -4,7 +4,7 @@ import (
 	"math/big"
 	"testing"
 
-	hu "github.com/ava-labs/subnet-evm/plugin/evm/orderbook/hubbleutils"
+	hu "github.com/ava-labs/subnet-evm/hubbleutils"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/assert"
 )
