+++ plugin/evm/orderbook/mocks.go
@@ -4,7 +4,7 @@ import (
 	"math/big"
 
 	"github.com/ava-labs/subnet-evm/core/types"
-	hu "github.com/ava-labs/subnet-evm/plugin/evm/orderbook/hubbleutils"
+	hu "github.com/ava-labs/subnet-evm/hubbleutils"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/mock"
 )
