+++ plugin/evm/orderbook/trading_apis.go
@@ -14,7 +14,7 @@ import (
 	"time"
 
 	"github.com/ava-labs/subnet-evm/eth"
-	hu "github.com/ava-labs/subnet-evm/plugin/evm/orderbook/hubbleutils"
+	hu "github.com/ava-labs/subnet-evm/hubbleutils"
 	"github.com/ava-labs/subnet-evm/rpc"
 	"github.com/ava-labs/subnet-evm/utils"
 	"github.com/ethereum/go-ethereum/common"
