+++ precompile/contracts/bibliophile/oracle.go
@@ -3,7 +3,7 @@ package bibliophile
 import (
 	"math/big"
 
-	hu "github.com/ava-labs/subnet-evm/plugin/evm/orderbook/hubbleutils"
+	hu "github.com/ava-labs/subnet-evm/hubbleutils"
 	"github.com/ava-labs/subnet-evm/precompile/contract"
 
 	"github.com/ethereum/go-ethereum/common"
