+++ precompile/contracts/bibliophile/orderbook.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/ava-labs/subnet-evm/precompile/contract"
 
-	hu "github.com/ava-labs/subnet-evm/plugin/evm/orderbook/hubbleutils"
+	hu "github.com/ava-labs/subnet-evm/hubbleutils"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
 )
