+++ precompile/contracts/juror/ioc_orders.go
@@ -4,8 +4,8 @@ import (
 	"errors"
 	"math/big"
 
+	hu "github.com/ava-labs/subnet-evm/hubbleutils"
 	ob "github.com/ava-labs/subnet-evm/plugin/evm/orderbook"
-	hu "github.com/ava-labs/subnet-evm/plugin/evm/orderbook/hubbleutils"
 	b "github.com/ava-labs/subnet-evm/precompile/contracts/bibliophile"
 	"github.com/ethereum/go-ethereum/common"
 )
