+++ precompile/contracts/juror/ioc_orders_test.go
@@ -5,8 +5,8 @@ import (
 	"math/big"
 	"testing"
 
+	hu "github.com/ava-labs/subnet-evm/hubbleutils"
 	"github.com/ava-labs/subnet-evm/plugin/evm/orderbook"
-	hu "github.com/ava-labs/subnet-evm/plugin/evm/orderbook/hubbleutils"
 	b "github.com/ava-labs/subnet-evm/precompile/contracts/bibliophile"
 	"github.com/ethereum/go-ethereum/common"
 	gomock "github.com/golang/mock/gomock"
