+++ precompile/contracts/jurorv2/matching_validation.go
@@ -5,8 +5,8 @@ import (
 	"math/big"
 	"strings"
 
+	hu "github.com/ava-labs/subnet-evm/hubbleutils"
 	ob "github.com/ava-labs/subnet-evm/plugin/evm/orderbook"
-	hu "github.com/ava-labs/subnet-evm/plugin/evm/orderbook/hubbleutils"
 	b "github.com/ava-labs/subnet-evm/precompile/contracts/bibliophile"
 	"github.com/ava-labs/subnet-evm/utils"
 	"github.com/ethereum/go-ethereum/common"
