+++ precompile/contracts/jurorv2/notional_position.go
@@ -1,7 +1,7 @@
 package jurorv2
 
 import (
-	hu "github.com/ava-labs/subnet-evm/plugin/evm/orderbook/hubbleutils"
+	hu "github.com/ava-labs/subnet-evm/hubbleutils"
 	b "github.com/ava-labs/subnet-evm/precompile/contracts/bibliophile"
 )
